Extend identity counter bounds with int64

+++ internal/data/verify_users.go
@@ -26,9 +26,9 @@ type VerifyUsers struct {
 	BirthDateUpperBound       sql.NullString `db:"birth_date_upper_bound"`
 	EventData                 sql.NullString `db:"event_data"`
 	ExpirationDateUpperBound  sql.NullString `db:"expiration_date_upper_bound"`
-	IdentityCounter           int32          `db:"identity_counter"`
-	IdentityCounterLowerBound int32          `db:"identity_counter_lower_bound"`
-	IdentityCounterUpperBound int32          `db:"identity_counter_upper_bound"`
+	IdentityCounter           int64          `db:"identity_counter"`
+	IdentityCounterLowerBound int64          `db:"identity_counter_lower_bound"`
+	IdentityCounterUpperBound int64          `db:"identity_counter_upper_bound"`
 	Selector                  int32          `db:"selector"`
 	TimestampLowerBound       sql.NullTime   `db:"timestamp_lower_bound"`
 	TimestampUpperBound       sql.NullTime   `db:"timestamp_upper_bound"`
